datatype/[un]pack_external[_size]: move subroutines down to ompi/datatype so it can be directly used by test/datatype/external32
